oysterpack/pyproject.toml
[tool.poetry]
name = "oysterpack"
version = "0.1.0"
description = ""
authors = ["oysterpack.inc <2zP6OaymGZhJTHFvQMzp6Ze>"]
readme = "README.md"
packages =[
    {include="oysterpack"}
]

[tool.poetry.dependencies]
python = "^3.11"
py-algorand-sdk = "^2.0.0"
python-ulid = "^1.1.0"
pyteal = "^0.23.0"
beaker-pyteal = "^1.0.0b2"
click = "^8.1.3"
click-shell = "^2.1"
sqlalchemy = "^2.0.0"
pynacl = "^1.5.0"
reactivex = "^4.0.4"
websockets = "^10.4"

[tool.poetry.group.dev.dependencies]
coverage = "^7.0.5"
black = "^22.12.0"
mypy = "^1"
flake8 = "^6.0.0"
pylint = "^2.16.2"
alembic = "^1.9.4"
pyright = "^1.1.296"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
